[ISIS-5] Restructure app
